Refactor away from single main.rs, bump version
