+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "colorbuddy"
-version = "0.1.6"
+name = "color_buddy"
+version = "1.0.0"
 authors = ["Adam Henley <adamazing@gmail.com>"]
 edition = "2021"
 description = "Generates a color palette based on an image."
@@ -8,6 +8,10 @@ repository = "https://github.com/adamazing/colorbuddy.git"
 license = "MIT"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[lib]
+name = "color_buddy"
+path = "src/lib.rs"
 
 [dependencies]
 anyhow = "1.0.65"
@@ -19,6 +23,7 @@ console = "0.15.5"
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.140"
 chrono = {version= "0.4.41", features = ["serde"]}
+thiserror = "1.0.0"
 
 [[bin]]
 name = "colorbuddy"
