+++ README.md
@@ -2,7 +2,7 @@
 
 <p align="center">  
   <a href="https://techforpalestine.org/learn-more"><img alt="Free Palestine 🍉🇵🇸" src="https://img.shields.io/badge/%F0%9F%87%B5%F0%9F%87%B8_Ceasefire_Now-techforpalestine.org-000000?labelColor=gray&color=D83838" /></a>
-  <img alt="Version" src="https://img.shields.io/badge/version-0.1.6-blue.svg?cacheSeconds=2592000" />
+  <img alt="Version" src="https://img.shields.io/badge/version-1.0.0-blue.svg?cacheSeconds=2592000" />
   <a href="#" target="_blank">
     <img alt="License: MIT" src="https://img.shields.io/badge/License-MIT-yellow.svg" />
   </a>
@@ -136,7 +136,7 @@ I will endeavour to respond in a relatively timely manner but provide no guarant
 ## Roadmap
 
 - [ ] Allow users to generate palette information used by their graphics tools/applications
-- [ ] Refactor and pay down technical debt
+- [x] ~~Refactor and pay down technical debt~~
 - [x] ~~Allow users to generate a separate standalone palette image~~
 - [x] ~~Allow users to specify multiple images upon which to apply the same options~~
 - [x] ~~Allow users to specify an output file/directory~~
@@ -144,7 +144,7 @@ I will endeavour to respond in a relatively timely manner but provide no guarant
 
 ## Author
 
-👨 **Adam Henley (he/him)**
+👨 **Adam Henley (they/them)**
 
 * Website: [https://adamhenley.com](https://adamhenley.com)
 * Twitter: [@adamofgreyskull](https://twitter.com/adamofgreyskull)
