+++ src/palette/mod.rs
@@ -1,0 +1,10 @@
+//! Color palette extraction functionality
+//!
+//! This module contains the core algorithms for extracting color palettes
+//! from images using various quantization methods.
+
+pub mod extractor;
+pub mod converter;
+
+// Re-export the main extraction function
+pub use extractor::extract_palette;
